farm/requirements.txt
azure-iot-device==2.13.0
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorzero==2.0
deprecation==2.1.0
evdev==1.7.0
future==1.0.0
gpiozero==2.0.1
grove.py==0.6
idna==3.7
interfaces==0.0.4
iso8601==2.1.0
janus==1.0.0
lgpio==0.2.2.0
mypy-extensions==1.0.0
packaging==24.0
paho-mqtt==1.6.1
pathspec==0.12.1
platformdirs==4.2.1
pynmea2==1.19.0
pyserial==3.5
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
requests-unixsocket==0.3.0
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
seeed-python-reterminal==0.5
serial==0.0.97
smbus2==0.4.3
typing_extensions==4.11.0
urllib3==1.26.18
